[input] basic keyboard support + [physics] bugfix for force processing
[physics] There was a bug that caused weird stuttering when multiple
forces were applied. Force is now applied and reduced more fluently.
Bonus: game_objects now generally have a physics instance and are now
influenced by gravity.

[input] Basic keyboard support added. Also the system now compares all
pressed buttons to a button mapping file which then says if "left" or
"right" or "attack" or "jump" is pressed. This allows other classes to
ask for "attack" action instead of "A" key or "Space". The player class
already uses the new system.