[input] more supported keyboard keys + [physics] added collisions
A physics object can now tell if it collides with any other object on
the map. For grounding checks the collision check can also check if a
certain point collides with any other object. I implemented a simple
ground check for this that populates a list of points due to density and
width settings and checks them for any collision. If grounded, gravity
shows no effect. For this setting to work completely, levels should get
their own physics controllers (or collision controllers) too.